Refactor: move config reducers to its own folder and namespace
