+++ client/src/ports/config/nightWatchConfigAdapters.js
@@ -1,7 +1,7 @@
 import NightWatchConfigEntity from '../../domain/NightWatchConfigEntity';
 
-export const toNightWatchConfigEntity = maxTotalTimeSpent =>
-  new NightWatchConfigEntity({ maxTotalTimeSpent });
+export const toNightWatchConfigEntity = nightWatchConfig =>
+  new NightWatchConfigEntity(nightWatchConfig);
 
 export const toPlainNightWatchConfigEntity = nightWatchConfigEntity => ({
   maxTotalTimeSpent: nightWatchConfigEntity.maxTotalTimeSpent,
