+++ client/src/store/actions/actions.js
@@ -3,7 +3,5 @@ export const CHARACTER_REMOVE = 'CHARACTER_REMOVE';
 export const CHARACTER_NAME_UPDATE = 'CHARACTER_NAME_UPDATE';
 export const CHARACTER_SLEEP_TIME_INCREASE = 'CHARACTER_SLEEP_TIME_INCREASE';
 
-export const MAX_TOTAL_TIME_SPENT_INCREASE = 'MAX_TOTAL_TIME_SPENT_INCREASE';
-
 export const ADD_NIGHT_WATCH_PLAN = 'ADD_NIGHT_WATCH_PLAN';
 export const FETCH_NIGHT_WATCH_PLAN_FAILED = 'FETCH_NIGHT_WATCH_PLAN_FAILED';
