+++ client/src/store/actions/config/actions.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line
+export const MAX_TOTAL_TIME_SPENT_INCREASE = 'MAX_TOTAL_TIME_SPENT_INCREASE';
