+++ client/src/store/reducers/config/selectors.js
@@ -1,0 +1,3 @@
+export const getConfig = ({ config }) => ({ ...config });
+
+export const getMaxTotalTimeSpent = ({ config: { maxTotalTimeSpent } }) => maxTotalTimeSpent;
